+++ config.json
@@ -1,9 +1,8 @@
 {
     "base_url": "http://127.0.0.1:32400",
-    "token": "",
+    "token": "q2Ja8qquBNgaPoPok7jj",
     "bulk_files": [
-        "bulk_import.txt",
-        "test.txt"
+        "bulk_import.txt"
     ],
     "tv_library": [
         "TV Shows"
