+++ main.py
@@ -9,7 +9,7 @@ from src.ui.cli import PlexPosterCLI
 
 
 # Interactive CLI mode flag
-INTERACTIVE_CLI = False  # Set to False when building executable for GUI default
+INTERACTIVE_CLI = True  # Set to False when building executable for GUI default
 
 
 def check_playwright_browsers():
