+++ src/ui/cli/__init__.py
@@ -1,0 +1,3 @@
+from .main import PlexPosterCLI
+
+__all__ = ['PlexPosterCLI']
