+++ src/ui/gui/__init__.py
@@ -1,0 +1,5 @@
+"""GUI package for Plex Poster Set Helper."""
+
+from .main import PlexPosterGUI
+
+__all__ = ["PlexPosterGUI"]
