+++ src/ui/gui/handlers/__init__.py
@@ -1,0 +1,6 @@
+"""Handler modules for business logic."""
+
+from .scrape_handler import ScrapeHandler
+from .label_handler import LabelHandler
+
+__all__ = ["ScrapeHandler", "LabelHandler"]
