+++ src/ui/gui/tabs/__init__.py
@@ -1,0 +1,15 @@
+"""Tab modules for the GUI."""
+
+from .poster_scrape_tab import PosterScrapeTab
+from .bulk_import_tab import BulkImportTab
+from .title_mappings_tab import TitleMappingsTab
+from .manage_labels_tab import ManageLabelsTab
+from .settings_tab import SettingsTab
+
+__all__ = [
+    "PosterScrapeTab",
+    "BulkImportTab",
+    "TitleMappingsTab",
+    "ManageLabelsTab",
+    "SettingsTab",
+]
