+++ src/ui/gui/widgets/__init__.py
@@ -1,0 +1,15 @@
+"""Widget helpers and utilities."""
+
+from .ui_helpers import UIHelpers
+from .entry_row import EntryRow
+from .double_entry_row import DoubleEntryRow
+from .slider_control import SliderControl
+from .dynamic_list import DynamicList
+
+__all__ = [
+    "UIHelpers",
+    "EntryRow",
+    "DoubleEntryRow",
+    "SliderControl",
+    "DynamicList",
+]
